requirements.txt
torch
torchvision
torchserve
onnx
onnxruntime
PyYAML==6.0.1
captum==0.6.0
numpy==1.21.3
opencv-python==4.5.4.58
Pillow==10.0.1